+++ libs/ra8_io/src/ra8_io_blockdev_mram.c
@@ -54,9 +54,9 @@ static const char* const s_tag = "ra8_io_blockdev_mram";
  * @since 0.1.0
  */
 typedef enum : uint32_t {
-  k_ra8_io_mram_erase_block_bytes = (uint32_t)k_ra8_mram_block_size_bytes,
+  k_ra8_io_mram_erase_block_bytes = k_ra8_mram_block_size_bytes,
   /**< MRAM erase/program unit in bytes (32). */
-  k_ra8_io_mram_program_bytes = (uint32_t)k_ra8_mram_write_size_bytes,
+  k_ra8_io_mram_program_bytes = k_ra8_mram_write_size_bytes,
   /**< MRAM minimum program granularity in bytes (32). */
   k_ra8_io_mram_erase_unit_blocks = 1U,
   /**< Logical-block erase granularity: 1 (MRAM erases in 32-byte sub-units). */
